[Abraham]:create components.
