+++ src/app/components/popupDynamic.tsx
@@ -1,0 +1,93 @@
+import Popup from "reactjs-popup";
+import { AiOutlineClose } from "react-icons/ai";
+import { useEffect, useState } from "react";
+
+const MOBILE_THRESHOLD = 768;
+
+export function PopupDynamic({ open, selectedItem, onClose }) {
+    const [isMobile, setIsMobile] = useState(open);
+    const [isModalOpen, setModalOpen] = useState(false);
+
+    const closeModal = () => {
+        setModalOpen(false);
+        onClose(); // Close modal in parent component
+    };
+
+    useEffect(() => {
+        setModalOpen(open);
+        setIsMobile(window.innerWidth < MOBILE_THRESHOLD);
+
+        const handleResize = () => {
+            setIsMobile(window.innerWidth < MOBILE_THRESHOLD);
+        };
+
+        handleResize();
+        window.addEventListener('resize', handleResize);
+
+        return () => {
+            window.removeEventListener('resize', handleResize);
+        };
+    }, [open]);
+
+    return (
+        <Popup
+            open={isModalOpen}
+            modal
+            nested
+            contentStyle={{
+                maxWidth: 'lg:max-w-screen-sm sm:max-w-screen-md', width: isMobile ? '80%' : '30%',
+            }}
+            closeOnDocumentClick={true}
+        >
+            <div className="modal text-right p-4">
+                <button
+                    style={{ color: 'gray', fontWeight: 'bold' }}
+                    className="close-button">
+                    <AiOutlineClose
+                        onClick={closeModal}
+                        cursor='pointer'
+                        className='text-black'
+                        size={20}
+                    />
+                </button>
+                <div className=" mx-auto text-center mb-5 p-2">
+                    <h1 className="text-xl text-black font-bold m-5">Address updated</h1>
+                    <h2 className="text-base text-black font-bold m-4 ">New address to your account</h2>
+                    <h2 className="mt-3   text-base">
+                        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+                        incididunt ut labore et dolore magna aliqua. Ut emin ad minim veniam, quis nortrud
+                        exercitation ulloamco.
+                    </h2>
+                    <h2 className="text-black text-base m-4">Nisi ut aliquip ex ea commodo consequat</h2>
+                    <br/>
+                    <br/>
+                    <br/>
+                    <br/>
+                    <br/>
+                </div>
+                <div className="footer text-center mt-5">
+                    <button type="button"
+                            className="
+                                     py-2.5
+                                     w-1/2
+                                     px-5 me-2
+                                     mb-2
+                                     text-sm
+                                     font-medium
+                                     bg-pink-300
+                                     text-white
+                                     focus:outline-none
+                                     rounded-full
+                                     border-gray-200
+                                     hover:bg-white
+                                     hover:border-pink-300
+                                     hover:text-pink-300
+                                  "
+                            onClick={closeModal}
+                    >UNDERSTOOD
+                    </button>
+                </div>
+            </div>
+        </Popup>
+    );
+}
